Refactor CSR
