+++ packages/client/src/csr.ts
@@ -1,32 +1,36 @@
-import { CsrPayload } from "decorator-shared/types";
 import { buildHtmlElement } from "./helpers/html-element-builder";
+import type { CsrPayload } from "decorator-shared/types";
 
-const findOrError = (id: string) => {
-    const el = document.getElementById(`decorator-${id}`);
-
-    if (!el) {
-        throw new Error(`No elem:${id}. See github.com/navikt/decorator-next`);
-    }
-
-    return el;
-};
-
-const hydrate = () =>
-    fetch(findOrError("env").dataset.src!)
-        .then((res) => res.json())
-        .then((elements: CsrPayload) => {
-            (["header", "footer"] as const).forEach(
-                (key) => (findOrError(key).outerHTML = elements[key]),
-            );
-            window.__DECORATOR_DATA__ = elements.data;
-
-            elements.scripts
-                .map(buildHtmlElement)
-                .forEach((script) => document.body.appendChild(script));
+export const fetchAndRenderClientSide = async (url: string) => {
+    const csrAssets = await fetch(url)
+        .then((res) => res.json() as Promise<CsrPayload>)
+        .catch((e) => {
+            console.error("Error fetching CSR assets: ", e);
+            return null;
         });
 
-if (document.readyState === "loading") {
-    document.addEventListener("DOMContentLoaded", hydrate);
-} else {
-    hydrate();
-}
+    if (!csrAssets) {
+        console.error("Failed to fetch CSR assets!");
+        return;
+    }
+
+    const headerEl = document.getElementById("decorator-header");
+    if (headerEl) {
+        headerEl.outerHTML = csrAssets.header;
+    }
+
+    const footerEl = document.getElementById("decorator-footer");
+    if (footerEl) {
+        footerEl.outerHTML = csrAssets.footer;
+    }
+
+    window.__DECORATOR_DATA__ = elements.data;
+
+    csrAssets.css.map(buildHtmlElement).forEach((cssElement) => {
+        document.head.appendChild(cssElement);
+    });
+
+    csrAssets.scripts.map(buildHtmlElement).forEach((scriptElement) => {
+        document.body.appendChild(scriptElement);
+    });
+};
